// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/employee-work-experience-details-model.ts
export class EmployeeWorkExperienceDetailsModel {
    employeeWorkExperienceId: string;
    employeeId: string;
    designationId: string;
    company: string;
    comments: string;
    fromDate: Date;
    toDate: Date;
    isArchived: boolean;
    operationsPerformedBy: string;
    firstName: string;
    surName: string;
    userName: string;
    email: string;
    designationName: string;
    timeStamp: any;
    totalCount: number;
}